requirements.txt
pycryptodome==3.15.0
bcrypt==4.0.1
